+++ packages/core/core/src/types.ts
@@ -205,13 +205,14 @@ export interface ProductGetters<PRODUCT, PRODUCT_FILTER> {
   getPrice: (product: PRODUCT) => AgnosticPrice;
   getGallery: (product: PRODUCT) => AgnosticMediaGalleryItem[];
   getCoverImage: (product: PRODUCT) => string;
-  getFiltered: (products: PRODUCT[], filters?: PRODUCT_FILTER) =>
-    PRODUCT[];
+  getFiltered: (products: PRODUCT[], filters?: PRODUCT_FILTER) => PRODUCT[];
   getAttributes: (products: PRODUCT[] | PRODUCT, filters?: Array<string>) => Record<string, AgnosticAttribute | string>;
   getDescription: (product: PRODUCT) => string;
   getCategoryIds: (product: PRODUCT) => string[];
   getId: (product: PRODUCT) => string;
   getFormattedPrice: (price: number) => string;
+  getTotalReviews: (product: PRODUCT) => number;
+  getAverageRating: (product: PRODUCT) => number;
   getBreadcrumbs?: (product: PRODUCT) => AgnosticBreadcrumb[];
   [getterName: string]: any;
 }
@@ -323,6 +324,7 @@ export interface AgnosticCategoryTree {
   slug?: string;
   items: AgnosticCategoryTree[];
   isCurrent: boolean;
+  count?: number;
   [x: string]: unknown;
 }
 
@@ -435,6 +437,8 @@ export interface FacetSearchResult<S> {
 
 export interface AgnosticFacetSearchParams {
   categorySlug?: string;
+  rootCatSlug?: string;
+  term?: string;
   page?: number;
   itemsPerPage?: number;
   sort?: string;