+++ packages/core/docs/contributing/changelog.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 2.0.9
+
+- added `getTotalReviews` and `getAverageRating` to `productGetters` ([#4958](https://github.com/DivanteLtd/vue-storefront/issues/4958))
+- fix ssr transitions
+
 ## 2.0.8
 
 - adds `getEmailAddress` getter in `userGetters` (#4939)